task1/requirements.txt
anyio
asgiref
Babel
beautifulsoup4
black
bootstrap4
certifi
cffi
charset-normalizer
click
colorama
coreapi
coreschema
cryptography
defusedxml
Delorean
Django
django-adminactions
django-allauth
django-bootstrap4
django-ckeditor
django-crispy-forms
django-extensions
django-filter
django-js-asset
django-model-utils
django-rest
django-rest-swagger
django-simple-history
django-tables2
django-tinymce4-lite
djangorestframework
djangorestframework-jwt
easy-thumbnails
h11
httpcore
httpx
humanize
idna
in-place
itypes
Jinja2
jsmin
Markdown
MarkupSafe
mimesis
mypy-extensions
mysql-connector-python
oauthlib
openapi-codec
packaging
pathspec
Pillow
platformdirs
prettytable
py-spy
pyarmor
pyarmor-webui
pyarmor.cli.core
pycparser
PyJWT
pyspellchecker
python-dateutil
python-decouple
python3-openid
pytz
pytz-deprecation-shim
requests
requests-oauthlib
retrying
rfc3986
sentry-sdk
simplejson
six
sniffio
soupsieve
sqlparse
tzdata
tzlocal
uritemplate
urllib3
uuid
wcwidth
wget
win-unicode-console
xlrd
xlwt
